Adds sync timeout for applications
Adds sync timeouts to several applications within the ArgoCD ApplicationSet.

This ensures that ArgoCD waits for the specified duration for each
application to sync before considering it as failed. The timeouts vary
based on the application's complexity and expected sync duration.
This avoids sync staying in limbo when there is a race condition and the presync hooks cannot even be created.